third_party/blink/web_tests/external/wpt/selection/anonymous/details-ancestor-expected.txt

This is a testharness.js-based test.
[FAIL] Selecting the default summary of <details> should report a DOM-visible ancestor
  assert_equals: expected "HTMLDetailsElement" but got "HTMLBodyElement"
Harness: the test ran to completion.
